Added Shopify, Twillio, Pusher tests & implemented changes to pass it
I added a test for Shopify, Twillio and Pusher to see that the package
allows to compute the process needed to check their request signatures.

To fit the specifications of the 3 of them, I had to add one parameter
more to the the Val type.

I also removed all code implemented to follwing the defined steps to
implement the package functionality and at the end it was deprecated
because it isn't used in the first, already, finished version.